Code review fixes and add Changelog entry.
