+++ CHANGELOG.md
@@ -3,9 +3,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 # Upcoming
 
+## StreamChat
 ### 🐞 Fixed
 - Fix crash in ListDatabaseObserver.startObserving() [#2177](https://github.com/GetStream/stream-chat-swift/pull/2177)
 - Make BaseOperation thread safe [#2198](https://github.com/GetStream/stream-chat-swift/pull/2198)
+
+## StreamChatUI
+### ✅ Added
+- Add channel list states; empty, error and loading views.
 
 # [4.19.0](https://github.com/GetStream/stream-chat-swift/releases/tag/4.19.0)
 _July 21, 2022_
